Extraer el transporte del simulador para que lo comparta la visión
Hay dos programas que publican telemetría en el puerto 2026: el
simulador y el sistema de visión real. El contrato les promete a los
equipos que pasan de uno al otro SIN TOCAR SU CÓDIGO, y esa promesa no
es solo sobre el formato del mensaje: también es sobre cómo se comporta
la conexión.

Cuánto se espera antes de descartar un mensaje que el cliente no leyó,
si a un cliente recién conectado se le manda enseguida o se espera al
siguiente ciclo, qué pasa cuando uno se cae. Todo eso es contractual
—está en las secciones 6.3 y 8 de CONTRATO.md— y con dos
implementaciones podría divergir sin que nadie lo note. Un equipo
probaría contra el simulador, le funcionaría, y el día de la competencia
se toparía con algo apenas distinto. Ese es el tipo de falla que aparece
en el peor momento y cuesta horas encontrar.

Así que ahora hay UNA sola implementación, en contrato/publicador.py, y
la usan los dos. mock_publisher.py pasa de 785 a 626 líneas.

Lo único que se agregó al extraerla es un parámetro `avisar`, que por
defecto imprime en consola —así el simulador se comporta igual que
antes— y le permite al sistema de visión mandar esos avisos a su panel.
Una pieza compartida no puede decidir por sus dos usuarios dónde se
imprimen las cosas.

contrato/ sigue sin dependencias, que era la condición para que esto no
fuera un retroceso: sockets e hilos son biblioteca estándar. Verificado
analizando los imports de los cuatro archivos de la carpeta.

Comportamiento medido antes y después del cambio, con test_client.py:
94 contra 93 mensajes en cinco segundos, 0 inválidos, 0 saltos de
secuencia, 19 ms de latencia media en los dos casos, y "contrato: OK,
sin errores" las dos veces. Probadas las dos formas de arrancarlo: como
script suelto desde adentro de contrato/, que es la que usan los
equipos, y como módulo desde la carpeta madre.

De paso se quitó un `import socket` que quedó sin uso.
